cukemerlin: isolate config file handling
And also put all configuration files in one temporary dir, instead of
temporary file per name.

This makes it possible to have one configuration file refering to
another, without needing magic variables and replacements in the test
case. It's also possible to add pid files and sockets in the same
isolated directory

This is possible by changing the working directory to a temporary
directory if the @config tag is available, and then use relative paths
for everything instead.

It is with this patch possible to generate naemon configuration files
too, where naemon.cfg points to object configuration files.

Resolves: MON-8813

Change-Id: I516d84073dd8da916c064de00068bce57a1eefa4
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>