cukemerlin: Allow async responses, and msg fail
Since some steps should spawn a socket in the future, and return success
when receiving data, we need to allow for asynchornous responses. Thus,
we need to be able to store a reference to the connection by the step,
and then say STEP_OK or STEP_FAIL later

This also changes the API for returning from steps, which makes it
possible to actually return a message when failing.

Precondition for MON-8700

Change-Id: Ied08d22bb0d92103356e2bdefdc6ece25b55a376
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>