cukemerlin: steps to match query handler result
To be able to determine the current status as the merlin module
interprets it, make cukemerlin able to ask and match result from the
query handler.

This adds the possibility to ask the query handler for information and
match a given set of rules. All rules needs to match at least one row
each.

Resolves MON-8700

Change-Id: I5548968f9dfeab56199b28eda198430af486c367
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>