cukemerlin: Copy only defined part of config_hash
It's the shortest length of available space and available data that
should be copied, not the longest length.

Using the longest length will result in one of two cases: Write outside
of allocated memory and overwrite some other vars, or read outside of
memory. (if hash length isn't equal to available space)

Thus, use the correct length

Change-Id: I69963cfcd8ad07962200a7ad35257d50a7f6cb7f
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>