Don't pretend to allow version differences
We don't care about merlin version number to provide backward
compatibility, so treat all version number differences as invalid
version.

Once we start provide backward compatibility, we should change this,
with handling of which compatibility modes needed to convert between
versions.

But until then, just treat version differences as errors.

This is also the first time we use default configured_peers, thus update
that to 0, for when we only talk IPC

Resolves: MON-8814

Change-Id: I8d9c6e9998c5b11b3958c8daa9452f9d08c52224
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>