Unify setting protocol version and merlin signature
All packets we send are funneled down to node_send_event(),
which does a lot of bookkeeping and things like that. Make sure
we set the signature and protocol version there, and also make
sure we do *ALL* bookkeeping there, instead of "almost all".

This patch looks dodgy, but is in fact safe, since all merlin
endpoints discard connections from clients that have the wrong
protocol or signature version anyway.

In preparation for MON-9012.

Change-Id: I25df9e9e9fdb221a8749be6f14264e76a0ef420f
Signed-off-by: Andreas Ericsson <ae@op5.se>