(temporarily) stop disconnecting nodes with bad object config
This change makes the "mon test dist" suite pass, which verifies
that we have at least the same tested functionality as we had the
last time we cut a release. Disconnecting peers and pollers with
a bad object config is a behaviour change, and the "mon test dist"
suite is not yet updated to reflect it.

Change-Id: Ie17b7c9aba1076498ba4a30b18144167abd79a81
Signed-off-by: Andreas Ericsson <ae@op5.se>