mon test dist: Fix variable usage when checking for config pushes
For some reason the csync_num_attempts variable doesn't seem to
be parsed properly when we read it from the query handler socket
as a #merlin nodeinfo request. I'm really not sure why, but this
seems to make all the tests pass while testing the same thing,
so we'll be just fine with that.

Change-Id: I419abd126bae5f02211f23ed3df04995ddf85823
Signed-off-by: Andreas Ericsson <ae@op5.se>