Fixup pointers in oconfsplit
This has been a problem for long, even though, by some reason, the
system worked anyway.

When creating temporary objects (hosts and hostgroups) when outputting
the poller configuration, it is really important not to modify the
existing running configuration, otherwise pointers will be invalid.

When migrating from lists to trees (well, linked lists to rbtrees), some
of those where overseen, since add_host_to_hostgroup and
add_parent_to_host modified not only the host/hostgroup, but also the
parent/member.

Since we only create an temporary instance which should be written to a
file, then removed, we should (ugh...) manually change the member list,
output to the file, and manually (ugh again...) remove the list.

This was seen first when changing to GTree, even though it still
existed earlier

Change-Id: If0ca1230bc68b68941ed5f2f0d4da0b82d75572c
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>