module: Re-enable disconnecting nodes that don't match
The unit tests expect that, but the integration tests do not.
Without this patch, the "mon test dist" suite runs just fine.
With this patch, the unit tests still fail due to some
misconceived notions of how Merlin actually works. This patch
is necessary in order to fix those misconceived notions in
the unit tests, however.

Change-Id: I92f4306fac72e49ff408bd3b5be18ff1a7be22ad
Signed-off-by: Andreas Ericsson <ae@op5.se>