tests: Make peer config sync tests easier to follow
We do that by starting with the normal case, where both peers
have the same hash for both info and expected as each other.
Each test can then decide which hash to send, which makes it
a bit more obvious what's supposed to happen.

Also, use the same hash for info and expected, since they
always are in Merlin where peers are concerned.

Change-Id: Ia651ccb9f4bea39fb2b6ac24b58fb030e246f0ea
Signed-off-by: Andreas Ericsson <ae@op5.se>