nebev2kvvec: Stop pretending we still know what rbtrees are
They were replaced with g_tree's a while back, but this one
piece of code still references them, causing build warnings.
Since the type is *actually* g_tree's, the code still works,
but the compiler warnings are ugly and distracting, so let's
get rid of them and also make the code correct again.

Change-Id: I632252becb8742717f176edd453186845d593533
Signed-off-by: Andreas Ericsson <ae@op5.se>