mon test dist: Symlink poller config instead of generating it
Otherwise the config hashes will never match, and the whole
suit will fail miserably.

While at it, make use of macros in the config so that all
master nodes create the same configuration for all pollers,
rather than having their name printed in the notification
commands, which causes the config to be different from the
expected and thereby prevents pollers from staying connected
to their masters.

Change-Id: I29b6e7479094a79b3161acffe63b2912a9675a6b
Signed-off-by: Andreas Ericsson <ae@op5.se>