module: Try connecting to unconnected nodes every once in a while
The "every once in a while" is whenever we send a pulse. It might
be better to schedule this closer, like once every 3-5 seconds,
but this avoids the problem with the old code, where one CPU got
pegged at 100% if a random node was killed off after all cluster
connections were finished.

Signed-off-by: Andreas Ericsson <ae@op5.se>