mon test dist: Stop sending downtime notifications
We don't use them for testing purposes, but they increase the
load on the systems where the tests run by exorbitant amounts,
since the notification scriptlet is kinda heavy to load and
runs a lot of times in quick succession when downtime is
rapidly scheduled, cancelled and expired.

With this patch, system load average stays well below 1, which
means we shouldn't have any issues running these tests in the
CI system.

Signed-off-by: Andreas Ericsson <ae@op5.se>