module: Use separate sockets for connect()ing and accept()ing
Since it's impossible to control the order in which the iobroker
processes events and it's equally possible that connect attempts
succeed to nodes that we at the same time accept inbound
connections from (where "at the same time" here is defined as
"in a single call to process iobroker events), it was all too
easy to end up mismanaging the sockets, leading to lost and
inappropriately closed sockets, which played merry hell with
the query handler.

This is fallout from fixing MON-9012.

Signed-off-by: Andreas Ericsson <ae@op5.se>