mon test dist: Use unique portnumber combinations for the mesh
Merlin uses listen_port + node->listen_port as sourceport
when connecting to nodes on the same host, since source
ip and, if needed, source port are used to identify nodes
in a merlin cluster. Incrementing port numbers by one means
we make it impossible for nodes to connect to each other
when there's more than three nodes in a single peer group.

This patch changes the portnumber incrementation algorithm
to take advantage of the fact that exponential increments
prevent "A + D" from ever being equal to "B + C" for any
given A, B, C and D that are distinct integers and either 1
or a power of the base.

The fact that A+D=D+A is irrelevant, even though both are
used, since it's enough if one node in that pair can connect
to the other. Since we only need this to be true for three
nodes or less, we use two as our base.

As a nice bonus, this makes the "mon test" testsuite run
about eight seconds quicker.

This patch is of a janitorial nature and as such lacks
relation to a particular issue.

Signed-off-by: Andreas Ericsson <ae@op5.se>