Add defaults for all(?) previously missing merlin events
Even the ones we may not want to use and that Merlin currently
doesn't handle.

The event defaults most likely need to be handmangled a little
when they want to be used, but at least one doesn't have to
spend 20 minutes looking up struct names and stuff when that
happens now.

This ties in with testing issues such as MON-8720 and its ilk.

Signed-off-by: Andreas Ericsson <ae@op5.se>