Send events to daemon before decoding
Otherwise the daemon can't read anything we send it and
the logs will be filled with messages like these:
  Nulling OOB ptr 3. cb: HOST_CHECK; type: 0; offset:
     0x2e06088; len: 568; overshot with 48258640 bytes
  CODEC: Failed to decode packet from 'ipc'. type: 7
     (HOST_CHECK); code: 0; len: 568

While at it, we remove the clunky triplicated code that filtered
out events the daemon doesn't know about and replace it with a
small helper, and also take the opportunity to correct some old
and invalid comments.

This is part of MON-9012.

Signed-off-by: Andreas Ericsson <ae@op5.se>