Documentation: Reformat and tidy up README
Added a explanatory section on what merlin actually is, and created headings for better compability with github.

Referenced Naemon as the default application to be used with.

Included the ascii-art from the output of the command "mon ascii merlin" as a simple logo.

Referenced op5.org as a source for additional information.

Signed-off-by: Fredrik Mikker <fmikker@op5.com>