Add mon oconf poller-fix
The idea, as suggested by peklof, is to remove configuration
files on pollers that shouldn't be there, by modifying the
naemon.cfg file and also removing files mentioned in cfg_file
statements. The only object config line we keep uncommented in
naemon.cfg is "cfg_dir=oconf", which is where the object config
file from the master (aptly named from-master.cfg) ends up.

While at it, we make sure to use the return value from the
function we call as the exit-code from mon.py.in, which
lets us call command-module functions directly from each
other while still retaining their return codes for the
benefit of the shell.

Tests are added to 'make check', but as their own target so
as to not interfere with the autogenerated actions for
'make check'.

This fixes issue MON-9259

Signed-off-by: Andreas Ericsson <ae@op5.se>
Signed-off-by: Sidharta Bhapur <sbhapur@op5.com>