Avoid pushing to pollers when some peers are disconnected
In a cluster with master1 and master2 being peers at the
top level and at least one poller, a change made on one of
the masters (let's go with master1, for simplicity) can
trigger a config sync to the poller before it syncs the
config to master2. In such cases, it's possible that master2
will push its (old) config to the poller before it receives
the updated configuration from master1. This leads to odd
behavior when restarting the cluster, and can make it take
quite a long time to do so.

In order to prevent that, this patch prevents masters from
pushing to pollers unless they have older config, but only
when not all peers are connected to each other.

Since peers can only be considered "active" when they have
matching configuration, that means all peers have to agree
on which config the pollers should run before they can
overwrite a newer config with an older one.

This patch also includes a refactoring so that the compat_cmp()
functions only get the information that they need, but so that
they get *ALL* the information that they need. Tests and fallout
is fixed to match the new world order.

This fixes MON-7969.

Signed-off-by: Andreas Ericsson <ae@op5.se>