janitor: Abort when unrecoverable errors occur
When we're left with no good option after a failed system call,
we would previously just ignore the situation and continue as
if nothing had happened, which would (most of the time) lead to
an enormous amount of more failed system calls.

Avoid that situation by exiting early in case it happens instead,
even though some of them can't happen at all on Linux....

On a related sidenote;
The "let's not allow developers to shut us up by casting to (void)"
change is perhaps the biggest mistake in the history of gcc.

Signed-off-by: Andreas Ericsson <ae@op5.se>