showlog: Use png icons instead of gif icons
The .gif icons don't exist in application/views/icons/x16/ (in ninja)
where they are assumed to be when we render the eventlog (in ninja).
This causes hostevent, serviceevent and flapping lines to show up
without an icon.

The .png icons do exist in that location, however, and we use them for
everything else already.

This fixes MON-9014.

Signed-off-by: Anton Lofgren <alofgren@op5.com>