oconf: Let rsync check file checksums
In some cases, configuration can be changed without affecting
modified date and size. If that happens, oconf push will not
push configuration to the poller even if the master's config is
different from the poller's config.

Adding the -c parameter to rsync tells rsync to compare checksums
which will catch this type of problem.

Verified manually by setting up a master and poller using instructions
at https://kb.op5.com/x/bID7 and following steps to reproduce.

I was unabel to get the poller in "INACTIVE" state, but I checked
the "change" timestamp on the poller's config after running
asmonitor mon oconf push poller01
on the master. Before adding the -c parameter, the Change
timestamp did not change. After adding the -c parameter, the
Change timestamp did change.

Manual inspection of the config file confirms that the contents of
the file was updated when running with the -c parameter and that
the contents of the file did not change when running without the
-c parameter.

This fixes MON-9391