Reduce blocking in main Naemon event loop
With the Merlin network IO now being completely handled inside the
Naemon event loop, its possible under high numbers of checks and TCP
congestion / full buffers that we could end up poll()'ing and blocking
for 100ms for each check Naemon processes.  This starves the even loop
for resources, Naemon cannot process check results, and we are not
reading from the Merlin sockets.  This then cascades to peers as they
being to block on a full TCP write buffer.

I've reduced the amount of time poll() will block for from 100ms to
10ms.  I wonder if we should not block at all in the event loop.