Cukemerlin: Notifcations for passive check results
To be able to deterministically tell how notifications are executed in a
distibuted environemnt due to external commands, add tests both for when
passive check commands is executed locally and over network.

To be consistent, the external command for a passive check result should
only be executed on the node hanlding the check. Doing so, it will be
simple to prove that the check notifies only once

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>