Mon test dist: Guarantee hosts are unreachable
Since a host that is a child of another hosts goes unreachable first on
next state change, we need to guarantee that all hosts are registered as
down or unreachable before submitting down states to get to unreachable.

Previously, mon test dist used the fact that all passive check results
were executed on the node recieving them, thus that all check results
were executed in the order they were entered. That's not a fact that can
be guaranteed in an asynchronous system. Thus, we need some other way of
syncing that. It's in fact a behaviour defined by naemon

Also, nowadays, the passive check result is registered on the node that
handles the object, thus, all nodes should notify, if they aren't
pollers that is configured not to notify. In that chase, they shouldn't

Make the MON-9729 test pass

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>