neamon callbacks: net_sec not closed during module deinit
net_sec wasn't closed and its iobroker callback not unregistered while
deinitializing the merlin module. This leads to naemon having a callback to a
pointer which eventually points into not accessible memory thus leading to a
segfault.

The fix is just to close and unregister the callback when the module is
deinitialized.

This probably fixes the segfault issue in MON-9655

Signed-off-by: Tomas Vestelind <tvestelind@op5.com>