cukemerlin: wait for daemons to exit when cleaning up
Wait for daemon to die after sending SIGTERM. If it hasn't died in 60 seconds,
send a SIGKILL. If it hasn't been cleaned up in 60 seconds after that then do
exit(1).

If SIGKILL doesn't work then something is probably wrong in the kernel
which is very unlikely. Having an exit(1) makes it explicit that all daemons
must stop before proceeding even if this case will probably never happen.

Signed-off-by: Tomas Vestelind <tvestelind@op5.com>