mon commands: Make mon start/stop/restart work on all platforms
We used to call them using the full path, but since we changed to
systemd unit this doesn't work on el7.

This change let systemd use the backward compabilitylayer to handle
the services instead.

This should be removed in the future to something more inteligent...

This is part of MON-9096

Signed-off-by: Mattias Ryrlén <mattiasr@op5.com>