Fixing Multiline parsing of Query Handler output
The existing parsing of Query Handler output expects a single line to
contain all the required data for it to be parsed correctly. We ran into
some instances where the output was split across multiple lines (short
reads?), causing parsing errors and or invalid output to checks such as the
distribution check. This change first rejoins all Query Handler output and
then splits on the rowsep.