Check execution: Track number of checks on each node
Now Merlin propagates the state of its objects when
they are updated. Due to this recent change we no longer
had tracking of the number of checks each host actually ran.

A fix to this issue was to update the number of checks
each node executed every time we receive a status update.
So if we for instance receive a status update for Service A
from peer1, then we know that peer1 is executing checks
for this service.

An intermission was also added to the active checks in montestdist.
The reason for this was that it appears that nodes didn't have time
to connect to each other causing the nodes to randomly take
responsibility to nodes.

This is part of MON-8720

Signed-off-by: Robin Engström <robin.engstrom@op5.com>