Check results: Reverting to sending check results
The idea of keeping the owning node responsible for
updating the state of its objects is good in theory.
However, the way merlin is currently implemented it
means a whole lot of extra maintenance with regards
to logging and notifications.

So this commit is a revert back to sending check results
instead of sending object updates from the owning node.

Signed-off-by: Robin Engström <robin.engstrom@op5.com>