hooks.c: store notifications for later sending
When handling a check result which generates a notification message the order
of which messages were sent over merlin was incorrect. The order previously was
to send the notification message in the middle of the check result processing
and then when the processing was done it would send the check result to other
nodes. The idea of the notification message is to let other nodes know the
current notification state of an object, but when a check result is sent right
after a notification all those stats were reset.

This is a "quick" fix where we store the notification messages until the next
sending of a check result. Since naemon works sequentially with regards to
sending notification then check result, we know that after a notification was
sent, the check result that caused the nofication would be sent. So we use this
fact and store the notification message, then when the check result is sent we
flush the stored notification, thus changing the order of the messages. This
means that the information registered when the notification message arrives
isn't overwritten by the check result.

This fixes MON-8720

Signed-off-by: Robin Engström <robin.engstrom@op5.com>