hooks.c: notification log messages fixed
When receiving a check result from a peer which generates a
NEB notification event, we will block if the peer is responsible
for the notification. However our logs would falsely state that
the notification originated from a poller when it came from a peer.

This fixes the logging in merlin where a peer handles their own
notification by stating correctly where the notification
originated from.

Part of MON-10042

Signed-off-by: Robin Engström <robin.engstrom@op5.com>