Tests: Cleanup of tests and logging removed
Removed logging of everytime we send a packet to avoid log cluttering.
Some optimizing of tests made as well.

Also fixed a bug where we would store a notification packet when
we received a check result from poller which would result in a
notification. But when a notification is triggered by a received
check result we do not send a check result directly after
which we do when we trigger a notification locally. So in these
cases we send the notification packet directly.

Moved some code around in hooks.c as well to get better structure
and overview.

Part of MON-10042

Signed-off-by: Robin Engström <robin.engstrom@op5.com>
Signed-off-by: op5 System <root@localhost>
Signed-off-by: Robin Engström <robin.engstrom@op5.com>
Signed-off-by: Robin Engström <engstrom.robin@gmail.com>