cukemerlin.c: Drop privileges before running tests
Cukemerlin is run as root, which in turn spawns instances of naemon as root.
After patching naemon to disallow running it as root user it caused
cukemerlin to fail because it wasn't able to spawn any naemon instances.

This patch implements drop_privileges() into cukemerlin which was previously
part of naemon. The function was removed from naemon when instead of dropping
privileges when starting naemon it instead doesn't allow us to start it as
root user.

Signed-off-by: Robin Engström <robin.engstrom@op5.com>