hooks.c: Send check result output with update
Previously we would parse the check result locally and
send data that had been stored within the service/host object
to peers and masters. We would also handle plugin output,
long plugin output and performance data in merlin on the remote
end instead of naemon.

When a check result generates a notification it uses the
output the check result and not from the object. So when
we need to notify instead of a poller as a master, we
would not have access to the long output as it would have
been stored in the object directly and wasn't available
for the notification procedure.

With this commit we instead set the plugin_output parameter
of the state object being sent to the check result output.
This output contains the whole output with standard output,
long output and performance data. On the receiving end
we then have the possibility to parse this data as we would
normally do with a locally generated check result.

Part of MON-9950

Signed-off-by: Robin Engström <robin.engstrom@op5.com>