test: Make binlog and stringutils tests able to fail
We always count the amount of failing tests, and print the failing
tests' output so that it is picked up by the tests. The exit code was
never adjusted though, so the build/CI environment never picked up on
errors. This patch makes sure we exit with a non-zero exit code if there
are failing tests.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>