oconf: Stop writing "poller-fix: We're not a poller.."
When I type
$ mon start
in a terminal, I get this output:

 poller-fix: We're not a poller, so not fixing object config
Starting merlind:                                          [  OK  ]
Starting naemon:                                           [  OK  ]

The first line is both weirdly indented, a message without a scope, and
violating "The rule of silence" [1], which dictates

    [...] well-behaved programs should treat their users' attention and
    concentration as being valuable and thus perform their tasks as
    unobtrusively as possible. That is, silence in itself is a virtue.

In short: users should not expect to find debug logging when not
providing --verbose, and users should definitively not get such
information on stdout.

This fixes MON-10070.

    1: http://www.linfo.org/rule_of_silence.html

Change-Id: I3ec39bc9ece052a6ad46be1cbab0a52e3c6baa00
Signed-off-by: Carl Helmertz <chelmertz@op5.com>