hooks.c: Send notification packet to owning poller
Notification packets are sent by default to peers and masters.
This is a problem when a master notifies instead of a poller.
If the poller doesn't receive the notification packet about
the object it's owning but not notifying for it will not be
"aware" that the notification was sent, thus not having a
value set for last_notification parameter. This is mainly
a cosmetic issue but could have unforseen issues in the future.

This fix sets the notification packet destination to include
all pollers that are configured for the same hostgroup as
the object being notified about is in. This will make all
pollers within the hostgroup to become "aware" of the
notification being sent.

Signed-off-by: Robin Engström <robin.engstrom@op5.com>