Add all host/service information to merlin_host/service_status object
For report_data to be registered correctly in the database the
merlin_host/service_status must contain all the data relevant to the object.
This reverts the changes where we would store check result data in the object
before sending to the receiving node.

This would cause the problem of the long output not appearing in notifications
to reemerge but instead of sending unparsed data as it were in the check result
we instead repack it on the receiving end upon reception and put it in a
check_result object. This means that we estimate how the check_result object
looked on the senders end as it were when they parsed it.

Part of MON-10142

Signed-off-by: Robin Engström <robin.engstrom@op5.com>