hooks.c: Use peer_id instead of id to determine object ownership
The merlin_node object has two identifying properties; id and peer_id.
The latter is used to determine ownership of objects while the other
is simply identification. So we must use peer_id to determine if
we should notify or not instead of the plain id.

Also; we were missing some count tracking of number of notifications
made by different relatives to us. Added tracking for pollers
and peers in a couple of cases.

Part of MON-10202

Signed-off-by: Robin Engström <robin.engstrom@op5.com>