Always set last notification time according to checking node
Instead of holding a packet containing the information about when the last
notification was sent, we use the fact that this information is contained in
every check result packet received.

The issue to begin with is that the check result will overwrite any timestamp
set to last_notification. However since the check result contains the information
we can simply forcefully set the last_notification value directly after the
check result has been processed.

Signed-off-by: Robin Engström <robin.engstrom@op5.com>