hooks.c: only block check results that were handled by another node
Previously we would block every check result from being propagated if
we ended up in a state where we'd processed a check result due to a
received packet. So we would assume that we'd end up in this state
due to a received check result from another node.

However, if we would for instance receive an external command with
a passive check result that we're responsible for we would again
end up in a state where we have a processed check result and it was
generated as a result of a received merlin event.

This patch makes sure we check what type of merlin event that was
received before deciding if we should block the packet or not.

Signed-off-by: Robin Engström <robin.engstrom@op5.com>