hooks.c: don't block passive check results from being propagated
Previously we assumed that if we had ended up in the NEBTYPE_SERVICECHECK_PROCESSED
as a result of receiving a merlin packet from another node that the packet would
contain a check result. And since we don't want check results to start bouncing
between nodes we would block it.

However, there is another way that we could end up in NEBTYPE_SERVICECHECK_PROCESSED
as a result from received packet. That is if a fellow node has sent us an external
command containing a passive check result for an object which turns out to be our
responsibility. In this case we still want to send the check result to our peers and
masters.

Part of MON-10221

Signed-off-by: Robin Engström <robin.engstrom@op5.com>