oconf.py.in: only restart wanted nodes
A bug was created when parallelizing oconf push which would cause all nodes
to be restarted when only one was selected. This would all nodes to be
restarted n-1 times where n is the number of nodes in a cluster when pushing config
via merlin.

This is a quick fix which was discovered by Benjamin Ritcey in support ticket:
https://jira.op5.com/browse/SUPPORT-23248

Part of MON-10309

Signed-off-by: Robin Engström <robin.engstrom@op5.com>