oconfsplit.c: sort servicegroup members
We want to make sure there is always a logical order of the members in a
servicegroup when writing the writing the poller configuration file as master.
Reason being that masters should always generate the "same" config with the same
hash values otherwise they will become unsynched and cause the poller to restart
indefinitely.

Part of MON-10336

Signed-off-by: Robin Engström <robin.engstrom@op5.com>