add tests for custom notifications and the takeover option
Custom notifications should always be handled by the owning node. No matter where
the command is sent since a cluster should be viewed as a distributed single system.
So we test that this is in fact the case.

Also; add tests for the takeover option with regards to notifications.
If a master has set its 'takeover' option to 'yes' then it should not only handle
a pollers check, but also its notification.

Part of MON-9971

Signed-off-by: Robin Engström <robin.engstrom@op5.com>