hooks.c: stop holding notification messages
Instead of holding off on sending a notification packet, we should assume
the if we enter the notification loop and we have a received packet in
the buffer, then that packet caused the notification to occur.

If that packet then contains a check result, we should be able to assume
that the sender has already notified. The check result will also contain
the last_notification time. At this point we can update our local object's
last_notification time with this time.

Signed-off-by: Robin Engström <robin.engstrom@op5.com>