Send custom notification when node has notifies=no
Prior to this commit, if a poller was setup with the setting notifies=no
and you would send a SEND_CUSTOM_HOST/SVC_NOTIFICATION to the Naemon
query handler, then no notification would be sent. This commit ensure
that the master correctly will send the notification.

Note that in the case we handle a custom notification on a node that
normally isn't responsible for the object, then we do not forward the
external command to any other nodes. This is to ensure that we only send
one notification in case the systems has disagreements about the
settings.

This also fixes an issue in general with notifications that might not be
sent in case of pollers with notifies=no.

This fixes MON-10239

Signed-off-by: Jacob Hansen <jhansen@op5.com>