Changing return from True to 0
As per the Python doc, passing True to sys.exit means exit code 1:

https://docs.python.org/2/library/sys.html

purge_naemon_log_files sends its exit code to cmd_purge which in
turn seems to give it to mon.

Changing this means that running commands that use this helper:

> mon log purge --remove-older-than=3m

...will no longer exit with code 1.

There may be better ways to resolve this if we necessarily want
this function to return True, but this at least seems to solve
the issue with the above command example never returning 0.