Use specific version of parallel_tests ruby gem
Previously no version of the parallel_tests ruby gem was specified. This
means that when the gem is installed it installs the latest version. In
the latest version of the gem the required version of ruby was up from
2.0.0 to 2.1.0. We have ruby 2.0.0 on our test environment. Thus gem
could not be installed and test fails.

In this commit we specifiy the latest version of the parallel_tests ruby
gem that requires ruby 2.0.0 so that tests will run again. This is
perhaps not a solution for the long run but it will do for now.

Signed-off-by: Daniel Nilsson <daniel.nilsson@op5.com>