Do not deregister NEB callback in post_config_init
Currectly deregistering a NEB in the same callback does not correctly
work in Naemon and results in a heap corruption (use-after-free). Whether
this should be allowed in Naemon could be up for discussion.

This commit changes the post_config_init function to not deregister the
callback at all. This can be safely done, due the callback only being
called during startup/shutdown, and due to the guards around the
NEBTYPE.

This fixes: MON-11365

Signed-off-by: Jacob Hansen <jhansen@op5.com>