Testcase for custom notification corrected
This commit fixes a previously erroneous testcase. The testcase should
ensure that only one peer master sends out a custom notification.
However the wrong host was defined in the test case (a host which that
peer would not be resposible for).

After the fix for MON-10239, this test case, would fail, due to the
error above.

Signed-off-by: Jacob Hansen <jhansen@op5.com>