Improve logic for custom notification handling
The logic for handling the custom notification external command had
become a little hard to follow. Futher to this, the logic in place did
not match the logic used for actually sending out notifications.

The result of this was a mismatch, that could cause custom notifications
to not be sent out correctly.

This commit should address all of the above.

Part of MON-11401

Signed-off-by: Jacob Hansen <jhansen@op5.com>