Binlog: store persistently over restarts
With this commit we now persistently store the binlog to file when
merlin shuts down, and read it again during startup. This ensures that
no events are missed if a node is restarted with disconnected nodes.

This functionality can be disabled by setting `binlog_persist = 0` in
the configuration file.

This fixes: MON-10734 & #29

Signed-off-by: Jacob Hansen <jhansen@op5.com>