Return exit code 1 if there are inactive nodes
Previously the 'mon node status' command would always exit with
exit code 0, regardless of whether there were any inactive
nodes in the listing or not.

This patch adds a new variable that is used to return exit
code 1 in the case where there is one or more inactive
nodes listed.