Fix segmentation fault on long log messages
In some very rare cases, very long log message might be passed to the
log function. If the resulting log message is longer than 4096, we would
segfault when accessing `msg[len]`, because the return value of
`vsnprintf` is "he number of characters that would have been written if
n had been sufficiently large"

To fix this, we set the len to 4095, in case the the message is too
long.

This fixes MON-12041

Signed-off-by: Jacob Hansen <jhansen@op5.com>