Apps: Print useful error when nodeinfo unavailable
When Merlin has a fatal error, for example because a poller has been
added with an invalid hostgroup, previously we would print a nasty
stacktrace when running `mon node status`, and a couple of more
functions.

This commit ensures we print a more friendly error message and point to
a relevant log file.

This fixes: MON-10954

Signed-off-by: Jacob Hansen <jhansen@itrsgroup.com>