Apps: Conditional start of op5-monitor service
In the current CI system, many components rely on the `mon restart` and
`mon start` commands. However the CI system, just install the required
packages (as defined by the rpm dependencies). This results in the
op5-monitor service being missing in quite a few CI tests, causing the
start/restart scripts to fail.

This commit therefore first checks if the op5-monitor service exists on
the system, before attempting to start it.

Signed-off-by: Jacob Hansen <jhansen@op5.com>