Implement public key encryption
This commit implements encyrption of Merlin communication on port 15551.
The implementation is done using libsodium and the encryption cipher
used is XSalsa20.

The encryption is done on the body of the merlin packet, the header
remains unencrypted.

With this commit the merlin packet header is increased, to accomendate
sending the nonce in the header. Due to this the
`MERLIN_PROTOCOL_VERSION` is increased.

This fixes MON-12110

Signed-off-by: Jacob Hansen <jhansen@itrsgroup.com>